+++ nuxt.config.ts
@@ -35,6 +35,12 @@ export default defineNuxtConfig({
       key: "",
       databaseId: "",
     },
+    r2: {
+      endpoint: "",
+      bucketName: "",
+      accessKeyId: "",
+      secretAccessKey: "",
+    },
     public: {
       umamiAnalytics: {
         hostUrl: "",
@@ -136,6 +142,11 @@ export default defineNuxtConfig({
     rateLimiter: {
       tokensPerInterval: 300, // Doubled limit to prevent unexpected rate limiting
     },
+    headers: {
+      contentSecurityPolicy: {
+        "img-src": ["'self'", "data:", "https://ik.imagekit.io"],
+      },
+    },
   },
 
   ogImage: {
@@ -190,8 +201,12 @@ export default defineNuxtConfig({
   },
 
   image: {
-    domains: ["cdn.warframestat.us", "wiki.warframe.com"],
+    domains: ["https://ik.imagekit.io"],
     dir: "assets/images",
+    format: ["avif", "webp"],
+    imagekit: {
+      baseURL: "https://ik.imagekit.io/redeemr/framedle",
+    },
   },
 
   nitro: {